Remove logic checking for an existing artifact ID
Currently, we check if an artifact with the given slug already exists in KV and
get its artifact ID if one does not exist. Once we migrate to D1, this won't be
so easy.

While we can always find a way to re-add this functionality later, I actually
don't think it's that important. As long as we don't delete the artifact ID from
the submission JSON once it's assigned, the ID won't ever change.