Remove superfluous data from JSON sent to D1
There is some redundant data that we were sending to KV that we don't want to
send to D1, because we don't want redundant data in the database.

However, if it turns out that we can't decode the multihash in submission-worker
because we can't use the dependency in the Workers runtime, we may have to
reassess.